server/requirements.txt
openai==1.97.1
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.7.4
starlette==0.47.2
google-generativeai==0.3.2
